+++ geohabnet/R/network_density.R
@@ -1,9 +1,11 @@
 #' Network density
 #'
 #' Calculates and plots the network density of a GeoNetwork object.
+#' The plot is produced as dispersal parameter vs edge density.
 #'
 #' @param x A GeoNetwork object
 #' @return Vector. Up to two ggplot2 objects
+#'
 #' @export
 setMethod("ndplot", signature = "GeoNetwork", function(x) {
   .ndplot(x)
@@ -33,19 +35,23 @@ setMethod("ndplot", signature = "GeoNetwork", function(x) {
   
   # Plot for betas
   bplot <- if (length(betas) > 0) {
-    ggplot2::ggplot(data.frame(x = unlist(betas), y = unlist(eds_beta)),
-                    ggplot2::aes(x, y)) +
+    xvals <- unlist(betas)
+    yvals <- unlist(eds_beta)
+    ggplot2::ggplot(data.frame(xvals = xvals, yvals = yvals),
+                    ggplot2::aes(xvals, yvals)) +
       ggplot2::geom_point(color = "red") +
       ggplot2::geom_line() +
       ggplot2::labs(x = "Betas", y = "Edge density") +
       ggplot2::ggtitle("Inverse powerlaw") +
       ggplot2::theme_bw()
   }
-  
+
   # Plot for gammas
   gplot <- if (length(gammas) > 0) {
-    ggplot2::ggplot(data.frame(x = unlist(gammas), y = unlist(eds_gamma)),
-                    ggplot2::aes(x, y)) +
+    xvals <- unlist(betas)
+    yvals <- unlist(eds_beta)
+    ggplot2::ggplot(data.frame(xvals = xvals, yvals = yvals),
+                    ggplot2::aes(xvals, yvals)) +
       ggplot2::geom_line() +
       ggplot2::geom_point(color = "green") +
       ggplot2::labs(x = "Gammas", y = "Edge density") +
@@ -54,7 +60,13 @@ setMethod("ndplot", signature = "GeoNetwork", function(x) {
   }
 
   if (!is.null(bplot) > 0 && !is.null(gplot)) {
-    bplot + gplot
+    patchwork::wrap_plots(bplot, gplot)
+  } else if (!is.null(bplot)) {
+    bplot
+  } else if (!is.null(gplot)) {
+    gplot
+  } else {
+    stop("No data to plot")
   }
 
 }