+++ geohabnet/R/sean.R
@@ -658,7 +658,7 @@ msean_onrast <- function(global = TRUE,
 #' Guy Blomme, Mónica Carvajal-Yepes, Danny L Coyne, Wilmer J Cuellar, Gregory A Forbes,
 #' Jan F Kreuze, Jürgen Kroschel, P Lava Kumar, James P Legg, Monica Parker, Elmar Schulte-Geldermann,
 #' Kalpana Sharma, Karen A Garrett,
-#' _Global Cropland .connectivity: A Risk Factor for Invasion and Saturation by Emerging Pathogens and Pests_,
+#' _Global Cropland connectivity: A Risk Factor for Invasion and Saturation by Emerging Pathogens and Pests_,
 #' BioScience, Volume 70, Issue 9, September 2020, Pages 744–758,
 #' \doi{10.1093/biosci/biaa067}
 #' @references Hijmans R (2023). _terra: Spatial Data Analysis_.